refactor(console): dictation controller in UI/Console_Modules (wave 1)
Extract ChatScreen's one-shot mic-button dictation lifecycle into
ConsoleDictationController (tldw_chatbook/UI/Console_Modules/dictation.py)
-- wave 1's proof of the "controller" collaborator kind (a plain object
owning state and behaviour with no region of its own), alongside tasks
3-4's region-widget proof (left/right rail).

Moved verbatim: the three module-level types (ConsoleDictationEvent,
ConsoleDictationLimitSignal, ConsoleStreamingDictationSession), their
private helpers (_INLINE_BREAK_COMMANDS, _join_segments,
_voice_command_chip_ack, the four _VOICE_ACK_* strings), and the 20
ChatScreen methods matching *dictation* whose bodies read or wrote
_console_dictation_state or one of its seven sibling attributes.

ChatScreen keeps six one-line delegations under their original names
(_request_console_dictation_start/_stop/_cancel,
_sync_console_dictation_availability, and the two @on-decorated message
handlers) for the entry points reached from outside this cluster, plus
eight read/write properties (_console_dictation_state and friends) so the
eleven non-dictation methods that read this state (hands-free wiring,
on_button_pressed, composer-action sync) needed zero body changes.

Every one of the 20 method bodies is a byte-for-byte copy (AST-verified
against the pre-extraction source): the controller's constructor binds
every external name those bodies reference -- framework services
(run_worker/post_message/set_timer/set_interval), app_instance, and named
accessors for the composer/chat-store/speak-status helpers -- under the
same names the original methods used, plus properties for the narrow,
explicitly-documented reach-back into the hands-free loop (a sibling
cluster this wave does not touch) and two not-yet-extracted composer/
workspace attributes.

Existing worker groups (console-dictation-start/-stop/-cancel) were left
as three distinct groups rather than unified into one "console-dictation"
group: they are already dictation-exclusive (never shared with other
screen collaborators), and merging their exclusive=True cancellation
pools would let a fresh start cancel an in-flight cancel's mic-release
worker -- a behaviour change the plan forbids.

Test-file monkeypatches of ChatScreen._create_console_dictation_session
(the only class-level patch target in the dictation test suite) still
need retargeting to ConsoleDictationController in a follow-up commit.
